feat: :sparkles: Add 'addNewBoard' action to boardSlice
Introduce a new action 'addNewBoard' in boardSlice to handle the addition of a new board with specified 'boardName' and 'boardCols'. The action creates a new board with a unique ID, sets it as inactive by default, and initializes 'boardTodos' based on the provided 'boardCols'. This enhances the functionality of the boardSlice to facilitate the creation of new boards in the application.